Fixes view controller context bug when interface orientation is changed from portrait to landscape when a full-screen view controller is present over the split view controller
Feeds the popover controller a fake view controller, so it releases the original (master), and does its dark magic.  Fixes bug when the master view controller is contextually wrong (the height of its top and bottom bars will be wrong) when a user launches app in portrait, opens a modal full-screen view controller in landscape, then closes the view controller in landscape.